ML-049/improve integratioon and move to folder
