+++ .rubocop.yml
@@ -61,7 +61,7 @@ RSpec/EmptyExampleGroup:
 
 Metrics/CyclomaticComplexity:
   Exclude:
-    - 'app/services/the_movie_db.rb'
+    - 'app/services/integration/the_movie_db.rb'
 Metrics/ClassLength:
   Exclude:
-    - 'app/services/the_movie_db.rb'
+    - 'app/services/integration/the_movie_db.rb'
