+++ lib/tasks/the_movie_db.rake
@@ -1,5 +1,5 @@
 desc 'Fill data base with TheMovieDB Api'
 task :themoviedb => :environment do
-  TheMovieDb.new.call
+  Integration::TheMovieDb.new.call
 end
